Enhanced app bars behavior
