+++ app/src/main/res/values/strings.xml
@@ -12,6 +12,9 @@
 
     <!-- Random Card Fragment -->
     <string name="save_card">Save</string>
+    <string name="cd_save_card">Save the card\'s image into app storage</string>
+    <string name="refresh">Refresh</string>
+    <string name="cd_refresh">Refresh the content with a new random card</string>
 
     <!-- navigation bar -->
     <string name="magic" translatable="false">Magic</string>
